+++ docs/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+## [0.5.0] - 2024-09-06
+
 ### Added
 
 -   Added a check for outdated packages
@@ -82,7 +84,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
-[Unreleased]: https://github.com/spapanik/phosphorus/compare/v0.4.1...main
+[Unreleased]: https://github.com/spapanik/phosphorus/compare/v0.5.0...main
+[0.5.0]: https://github.com/spapanik/phosphorus/compare/v0.4.1...v0.5.0
 [0.4.1]: https://github.com/spapanik/phosphorus/compare/v0.4.0...v0.4.1
 [0.4.0]: https://github.com/spapanik/phosphorus/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/spapanik/phosphorus/compare/v0.2.1...v0.3.0